Fix constants and units bug in ic1ion phys prop
